Address CodeRabbit review feedback
- maplibre-raster: revoke retained local-bytes blob URLs on raster removal
  so File-backed rasters no longer leak memory for the session.
- raster-export/ProcessingDialog: extract the duplicated fetchableUrl helper
  into a shared url-utils module.
- LayerPanel: clear any existing status timer before setting the raster
  export status, matching handleExportLayer.
- ProcessingDialog: document the WASM raster_in path fallback and the clear
  error the runner surfaces for non-fetchable inputs.
